+++ monstr.py
@@ -1,12 +1,13 @@
+import numpy
+
 from copy import copy
+from os import getcwd
 
 from support import *
-from os import getcwd
 from pathfinding.core.diagonal_movement import DiagonalMovement
 from pathfinding.core.grid import Grid
 from pathfinding.finder.a_star import AStarFinder
 from settings import CELL_SIZE
-import numpy
 
 
 class PathFinder:
