+++ requirments.txt
@@ -1,0 +1,3 @@
+pygame
+pathfinding
+numpy
