+++ settings.py
@@ -1,3 +1,5 @@
 CELL_SIZE = 100
 SCREEN_WIDTH = 2048
 SCREEN_HEIGHT = 1024
+LAB_SIZE = 100
+OFFSETT = 2
